Hospital Insurance BE node into 3 parts
